third_party/blink/web_tests/http/tests/security/agent-equality.html

<!DOCTYPE html>
<title>Agent Equality Test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
// This test checks whether the JavaScript agent is created in a proper
// granularity. See:
// https://html.spec.whatwg.org/multipage/webappapis.html#integration-with-the-javascript-agent-formalism
// For details on each test, see the comments of the iframe file.

function runTestCase(filename,testFunction,description) {
  promise_test(t => new Promise(resolve => {
    let iframe = document.createElement('iframe');
    t.add_cleanup(async () => iframe.remove());
    iframe.src = 'http://127.0.0.1:8000/security/resources/' + filename;

    window.addEventListener('message', evt => {
      const sourceFile = evt.data[0].substring(evt.data[0].lastIndexOf('/') + 1);
      if (sourceFile != filename) {
        return;
      }
      testFunction(evt);
      resolve();
    });

    document.body.appendChild(iframe);
  }), description);
}

runTestCase('agent-equality-same-origin.html',evt => {
  assert_equals(evt.data.length, 4);
  assert_equals(evt.data[1], evt.data[2]);
  assert_equals(evt.data[1], evt.data[3]);
}, 'Same-origin documents must share the same agent.');

runTestCase('agent-equality-cross-origin.html',evt => {
  assert_equals(evt.data.length, 3);
  assert_not_equals(evt.data[1], evt.data[2]);
},'Cross-origin documents should receive a different agent.');

runTestCase('agent-equality-skip-level-same-origin.html',evt => {
  assert_equals(evt.data.length, 3);
  assert_equals(evt.data[1], evt.data[2]);
},'Same-origin documents must share the same agent even if opened by cross-origin document.');

// This test got more complicated by default-enabling origin-agent clusters:
// If Origin-Agent clusters are default disabled - i.e. legacy behaviour -
// then frames on the same domain but different port should be assigned to the
// same agent cluster, because document.domain setting might make them
// quasi same-origin with synchronous access between them. But with OAC
// default enabled, this is no longer possible and they should receive
// different clusters. So the desired outcome of this test depends on whether
// origin-agent clusters are default enabled.
//
// We solve this by querying for default enabling first, and then
// assert_equals or assert_not_equals based on it. We use the origin agent
// cluster state of the main document - which does not request OAC - as a
// proxy for whether OAC is default enabled or not.
const is_the_main_document_origin_agent_clustered = window.originAgentCluster;
runTestCase('agent-equality-same-origin-different-ports.html',evt => {
  assert_equals(evt.data.length, 3);
  if (!is_the_main_document_origin_agent_clustered) {
    assert_equals(evt.data[1], evt.data[2]);
  } else {
    assert_not_equals(evt.data[1], evt.data[2]);
  }
},'Documents on same-origin-but-different-ports should receive the same agent (if default OAC disabled).');

runTestCase('agent-equality-different-schemes.html',evt => {
  assert_equals(evt.data.length, 3);
  assert_not_equals(evt.data[1], evt.data[2]);
},'Documents with different schemes (HTTP and HTTPS) should receive different agents.');

runTestCase('agent-equality-data-url.html',evt => {
  assert_equals(evt.data.length, 3);
  assert_not_equals(evt.data[1], evt.data[2]);
},'Frame loaded as data: URL should receive a unique agent that is different from the parent\'s.');

runTestCase('agent-equality-srcdoc.html',evt => {
  assert_equals(evt.data.length, 3);
  assert_equals(evt.data[1], evt.data[2]);
},'srcdoc iframe should receive the same agent as the parent\'s.');

runTestCase('agent-equality-after-detach.html',evt => {
  assert_equals(evt.data.length, 3);
  assert_equals(evt.data[1], evt.data[2]);
},'Same-origin documents must share the same agent even after detaching.');
</script>
